requirements.txt
pytest
antlr4-python3-runtime==4.13.0
fastapi
uvicorn[standard]